Reorganize into installable package
