+++ pumanoise/__init__.py
@@ -1,0 +1,4 @@
+__author__ = """Anze Slosar"""
+__email__ = ''
+
+from .pumanoise import *
