+++ scripts/MWComparison.py
