+++ setup.py
@@ -1,0 +1,11 @@
+from setuptools import setup
+
+setup(
+    name='PUMAnoise',
+#    version='0.1',
+    description='Noise calculator for PUMA telescope',
+    author='Anze Slosar',
+    author_email='anze@bnl.gov',
+    packages=['pumanoise'],
+    package_dir={'pumanoise':'pumanoise'},
+)
